Split backend guest auth package
